Split Model Server to several classes. Delete ModelServer.
